requirements.txt
flask==3.1.1
sentence-transformers==4.1.0
pypdf==5.6.1
python-dotenv==1.1.0

langchain==0.3.26
langchain-core==0.3.83
langchain-community==0.3.26
langchain-text-splitters==0.3.11

langchain-pinecone==0.2.8
langchain-groq

-e .
